# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Digital Health Pass API",
    "description": "To view the available API endpoints, first login via the /users/login API and then refresh this page in your browser.",
    "version": "1.0.0"
  },
...
  "servers": [
    {
      "url": "/api/v1/credentials-issuer",
      "description": "Sandbox"
    }
  ],
  "paths": {
...
    "/schemas": {
      "get": {
        "tags": [
          "schemas"
        ],
        "summary": "Get all schemas",
        "description": "Get all schemas with optional query parameters",
        "operationId": "querySchemas",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-hpass-issuer-id",
            "description": "Issuer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully queried schemas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    },
                    "payload": {
                      "type": "array",
                      "items": {
                        "type": "object"
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "post": {
        "tags": [
          "schemas"
        ],
        "summary": "Create schema",
        "description": "Create a new schema",
        "operationId": "createSchema",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-hpass-issuer-id",
            "description": "Issuer ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSchemaRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchemaInfoResponse"
                }
              }
            }
          },
          "400": {
            "description": "Schema in request body is invalid",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "errors": {
                          "type": "array",
                          "items": {
                            "type": "object"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/issuers": {
      "post": {
        "tags": [
          "issuers"
        ],
        "summary": "Onboard issuer",
        "description": "Onboard a new issuer",
        "operationId": "createIssuer",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
        ],
        "requestBody": {
          "description": "Issuer information",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssuerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "issuer onboarded Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
...
    "/issuers/{did}/metadata": {
      "get": {
        "tags": [
          "issuers"
        ],
        "summary": "Get issuer metadata",
        "description": "Get an issuer's metadata",
        "operationId": "getIssuerMetadata",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "did",
            "description": "Issuer DID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved issuer metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IssuerMetadataResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      },
      "put": {
        "tags": [
          "issuers"
        ],
        "summary": "Update issuer metadata",
        "description": "Update an issuer's metadata",
        "operationId": "updateIssuerMetadata",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "did",
            "description": "Issuer DID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Update metadata of an issuer",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully updated issuer metadata",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/issuers/public-keys": {
      "post": {
        "tags": [
          "issuers"
        ],
        "summary": "Rotate issuer keys",
        "description": "Rotate an issuer's public keys",
        "operationId": "rotateIssuerKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "description": "Rotate issuer keys",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "issuerId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully rotated issuer's keys",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/InvalidRequestError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SchemaInfoResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "payload": {
            "type": "object",
            "properties": {
              "@type": {
                "type": "string"
              },
              "modelVersion": {
                "type": "string"
              },
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "author": {
                "type": "string"
              },
              "authorName": {
                "type": "string"
              },
              "authored": {
                "type": "string"
              },
              "schema": {
                "type": "object",
                "properties": {
                  "$schema": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              },
              "proof": {
                "type": "object",
                "properties": {
                  "created": {
                    "type": "string"
                  },
                  "creator": {
                    "type": "string"
                  },
                  "nonce": {
                    "type": "string"
                  },
                  "signatureValue": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      },
      "CreateSchemaRequest": {
        "type": "object",
        "required": [
          "id",
          "name",
          "version",
          "schema"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "version": {
            "type": "string"
          },
          "schema": {
            "type": "object"
          }
        }
      },
      "CreateIssuerRequest": {
        "type": "object",
        "required": [
          "issuerId",
          "type"
        ],
        "properties": {
          "issuerId": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "did:dhp",
              "did:web'"
            ]
          },
          "private_key_cert": {
            "type": "string"
          },
          "public_key_cert": {
            "type": "string"
          }
        }
      },
      "CommonResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "payload": {
            "type": "object",
            "properties": {
              "ok": {
                "type": "boolean"
              },
              "id": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "IssuerMetadataResponse": {
        "type": "object",
        "properties": {
          "payload": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              }
            }
          }
        }
      },
      "BasicErrorMessage": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "responses": {
      "InvalidRequestError": {
        "description": "Missing required fields in request body",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Logged in user is unauthorized to make this request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Unexpected server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/BasicErrorMessage"
            }
          }
        }
      }
    }
  }
}